typing password component props and tidy up register types
