+++ Client/src/pages/signIn/SignIn.tsx
@@ -1,5 +1,5 @@
 import { zodResolver } from "@hookform/resolvers/zod";
-import { loginSchema, loginTypes } from "../../utils/schemas";
+import { loginSchema } from "../../utils/schemas";
 import { useForm, SubmitHandler } from "react-hook-form";
 import { signIn } from "../../utils/Posts";
 import Password from "../../components/Input/Password";
@@ -20,7 +20,7 @@ export default function SignIn() {
     resolver: zodResolver(loginSchema),
   });
 
-  const onSubmit: SubmitHandler<loginTypes> = async (data) => {
+  const onSubmit: SubmitHandler<SignInType> = async (data) => {
     try {
       const response = await signIn(data);
       if (response) {
@@ -71,22 +71,6 @@ export default function SignIn() {
                 <p className="error-message">{errors.email?.message}</p>
               )}
             </div>
-            {/* <div className="input">
-              <input
-                className={
-                  errors?.password
-                    ? "wrapper__text-input error"
-                    : "wrapper__text-input"
-                }
-                type="password"
-                placeholder="password"
-                {...register("password")}
-                name="password"
-              />
-              {errors?.password && (
-                <p className="error-message">{errors.password?.message}</p>
-              )}
-            </div> */}
             <Password
               className={classNameText}
               errors={errors}
