+++ Client/src/utils/Posts.ts
@@ -1,17 +1,15 @@
 import axios, {isAxiosError} from "axios";
-import { UserPostResponseTypes} from "../types/registrationTypes";
-import { loginTypes } from "./schemas";
 import { apiURL, authURL} from "./links";
 
 
-export async function signIn(data: loginTypes){
+export async function signIn(data: SignInType){
     try {
     const response = await axios
       .post(authURL+ 'local', {
         identifier: data.email,
         password: data.password,
       });
-    return response.data as UserPostResponseTypes;
+    return response.data as UserPostResponseType;
   } catch (error) {
     if (isAxiosError(error)) {
       if (error.response?.status === 400) {
@@ -33,7 +31,7 @@ export const postUser = async (data: CreateUserType) => {
     const response = await axios.post(apiURL + "user/post", data);
     localStorage.setItem("jwt", response.data.jwt);
     setAuthToken(response.data.jwt);
-    return response.data as UserPostResponseTypes;
+    return response.data as UserPostResponseType;
   } catch (err) {
     if (isAxiosError(err)) {
       if (err.response?.status === 400) {
@@ -46,13 +44,3 @@ export const postUser = async (data: CreateUserType) => {
     }
   }
 };
-
-// export function postPayment(endpoint: string, data: paymentsTypes){
-//   const response = axios.post(apiURL + endpoint, data).then(response=>response.data.data as paymentsResponseTypes).catch(err=>{throw new Error("Wystąpił nieoczekiwany błąd:\n"+err)});
-//   return response;
-// }
-
-// export function postProfile(endpoint: string, data: profileTypes){
-//   const response = axios.post(apiURL + endpoint, data).then(response=>response.data.data).catch(err=>{throw new Error("Wystąpił nieoczekiwany błąd:\n"+err)});
-//   return response;
-// }
